Fix endless loop in splitSystemUseEntries
Fix endless loop in splitSystemUseEntries using all available memory.
The problem happens on an iso where the system use was set to.
00 00 00 00 0d 55 58 41 00 00 00 00 00 00

This seems to be a CD-XA System Use Information.
The patch tests if the size of the SystemUseEntry is at least 3 bytes as
defined in SUSP112 and if not an empty List of SystemUseEntry is
returned.

Added a unit test that handles the case resulting in an endless loop.